netplay: extract `syncDebug*` to a separate source file
